% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biOmics.R
\docType{package}
\name{biOmics}
\alias{biOmics}
\alias{biOmics-package}
\title{biOmics}
\description{
biOmics allows you to Download data of samples from TCGA, ENCODE, ROADMAP,
using a ontology-based algorithm
}
\details{
The functions you're likely to need from \pkg{biOmics} are
\code{\link{biOmicsApp}}, \code{\link{geoDownloaderLinks}} and
\code{\link{filterRmapData}}. Otherwise refer to the vignettes to see
how to format the documentation.
}

