% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/search.R
\name{biomics.search}
\alias{biomics.search}
\title{Searches a term in TCGA, ENCODE, ROADMAP}
\usage{
biomics.search(term, experiment = "all")
}
\arguments{
\item{term}{Term to be searched}
}
\description{
Search a term using an ontology in the TCGA, ENCODE, ROADMAP databases
}

