% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/global.R
\name{biOmicsApp}
\alias{biOmicsApp}
\title{biOmics interface}
\usage{
biOmicsApp()
}
\value{
Open a connection to shiny
}
\description{
Calls UI interface
}
\examples{
biOmicsApp()
}
\seealso{
\url{https://www.encodeproject.org/search/}

\url{https://www.encodeproject.org/help/rest-api/}
}

