% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/download.R
\name{biOmicsDownload}
\alias{biOmicsDownload}
\title{biOmics.download}
\usage{
biOmicsDownload(lines, enc.file.type = NULL, rmap.file.type = NULL)
}
\arguments{
\item{lines}{biOmics.search output}

\item{enc.file.type}{Extension to be downloaed from encode database}

\item{rmap.file.type}{Extension to be downloaed from roadmap database}
}
\description{
Download data selected using the biOmics.search function
}
\seealso{
biOmics.search
}

