% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{biOmicsSearch}
\alias{biOmicsSearch}
\title{Searches a term in TCGA, ENCODE, ROADMAP}
\usage{
biOmicsSearch(term, experiment = NULL, report = FALSE,
  dir.report = "searchSummary")
}
\arguments{
\item{term}{Term to be searched. Example: 'brain' 'u87' etc.}

\item{experiment}{Experiment type}

\item{report}{Create a summary plot of the result found? Deafult: TRUE}

\item{dir.report}{Directory to save the html report. Default: "searchSummary"
\tabular{llll}{
Microarray \tab MiRNAMicroArray \tab RRBS \tab DNAsequencing\cr
ExpressionArray \tab Firehose \tab ChipSeq \tab fiveC \cr
ExonArray \tab DNAMethylation  \tab MRESeq  \tab RepliSeq \cr
RNASeq \tab miRNASeq \tab Rampage \tab Others
}}
}
\value{
A dataframe with the results of the query if it
        was successful
}
\description{
Search a term using an ontology in the TCGA, ENCODE, ROADMAP databases
}
\examples{
inst/examples/biomicsSearch.R
}

