% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/biOmics.R
\docType{data}
\name{biosample.roadmap}
\alias{biosample.roadmap}
\title{Terms of roadmap mapped automatically using ontology algorithm,
plus manual mapping for terms not found}
\format{A data frame with 437 rows and 3 variables}
\description{
Terms of roadmap mapped automatically using ontology algorithm,
plus manual mapping for terms not found
}
\keyword{internal}

