% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methylation.R
\name{calculate.pvalues}
\alias{calculate.pvalues}
\title{Calculate pvalues}
\usage{
calculate.pvalues(values, idx1, idx2, paired = TRUE, exact = TRUE,
  mc.cores = parallel::detectCores())
}
\arguments{
\item{paired}{Do a paired wilcoxon test? Default: True}

\item{matrix1}{Matrix to be compared with matrix 2}

\item{matrix2}{Matrix to be compared with matrix 1}
}
\value{
Methylation table
}
\description{
Calculate pvalues
}
\details{
Input: two matrix to be compared with wilcoxon test,
          a boolean value to do a paired or non-paired test
   Output: p-values (non-adj/adj) histograms, p-values (non-adj/adj)
}

