% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoprint.R
\name{create.oncoprint}
\alias{create.oncoprint}
\title{Creating a oncoprint}
\usage{
create.oncoprint(mut, genes, filename, color)
}
\arguments{
\item{mut}{Mutation file}

\item{genes}{Gene list}

\item{filename}{name of the pdf}

\item{color}{for the graph}
}
\description{
Creating a oncoprint
}
\examples{
mut <- TCGAbiolinks::TCGAquery_maf(tumor = "GBM", archive.name = "ucsc.edu_GBM.IlluminaGA_DNASeq_automated.Level_2.1.1.0")
create.oncoprint(mut=mut,genes=mut$Hugo_Symbol[1:10])
}

