% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methylation.R
\name{diffmean.prim.rec}
\alias{diffmean.prim.rec}
\title{calculate diffmean methylation between primary and recurent}
\usage{
diffmean.prim.rec(data)
}
\arguments{
\item{data}{Data frame probes vs patient}
}
\description{
calculate diffmean methylation between primary and recurent
   samples
}

