% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/divisors.R
\name{divisor}
\alias{divisor}
\title{Return list with all divisors of the number}
\usage{
divisor(number)
}
\arguments{
\item{x}{A number}
}
\value{
vector whit \code{x} list of divisors
}
\description{
\code{prime} The function return a array list with all divisors of the number
}
\examples{
divisor(72)
divisor(50)
\dontrun{
   divisor('a')
}
}
\seealso{
\url{http://en.wikipedia.org/wiki/Divisor}
}

