% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/roadmap.R
\name{eGeo}
\alias{eGeo}
\title{Create query from UI to be used with esearch - not completed
Create query from UI to be used with esearch - not completed}
\usage{
eGeo(...)
}
\arguments{
\item{List}{of parameters parameter = (Term,Field,Connector)
First parameter should contain the terms,
Field NULL which means all fields and connector NULL
For roadmap one parameter should be c ("roadmap epigenomics","Project","AND"),
For samples: c ("gsm","Filter","AND") )}
}
\description{
Create query from UI to be used with esearch - not completed
Create query from UI to be used with esearch - not completed
}
\examples{
\dontrun{
 eGeo(
   c ("h1 cell RRBS",NULL,NULL),
   c ("roadmap epigenomics","Project","AND"),
   c ("gsm","Filter","AND") )
}
}
\keyword{internal}

