% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/biOmics.R
\docType{data}
\name{encode.db}
\alias{encode.db}
\title{Roadmap metadata database.
The data set contains the following fields:
\itemize{
\item accession Acession name
\item  biosample biosample name
\item assay Assay name
\item lab lab name
\item target target name
\item description description name
\item organism organim name
}}
\format{A data frame with 4598 rows and 7 variables}
\description{
Roadmap metadata database.
The data set contains the following fields:
\itemize{
\item accession Acession name
\item  biosample biosample name
\item assay Assay name
\item lab lab name
\item target target name
\item description description name
\item organism organim name
}
}
\keyword{internal}

