% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/download.R
\name{encode.download}
\alias{encode.download}
\title{Download encode data}
\usage{
encode.download(lines, type = NULL, path = ".")
}
\arguments{
\item{lines}{encode.search output}

\item{type}{extesion of files to be downloaded}

\item{path}{Folder to save the file}
}
\description{
Download encode data previously selected using the encode.search function
}

