% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/download.R
\name{encodeDownload}
\alias{encodeDownload}
\title{Download encode data}
\usage{
encodeDownload(lines, type = NULL, path = ".")
}
\arguments{
\item{lines}{encode.search output}

\item{type}{extesion of files to be downloaded}

\item{path}{Folder to save the file}
}
\value{
Download encode data into path
}
\description{
Download encode data selected using the encodeSearch
}

