% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/encode.R
\name{encodeDownloader}
\alias{encodeDownloader}
\title{Download data from ENCODDE project}
\usage{
encodeDownloader(iSearch, iType, iTarget, iSample, iFType, iAssay, iAssembly,
  iOut)
}
\arguments{
\item{iSearch}{- bone chip}

\item{iType}{-   'experiment' 'publication' 'software' ''antibody_lot' 'web'}

\item{iTarget}{- 'transcription factor' 'RNA binding protein'
'tag' 'histone modification'}

\item{iSample}{- 'tissue' 'primary cell'}

\item{iFType}{- 'bam' 'bigWig' 'bed_broadPeak' 'broadPeak' 'fastq'}

\item{iAssay}{- 'ChIP-seq' 'RIP-chip' 'Repli-chip'}

\item{iAssembly}{- 'hg19' 'mm9'}

\item{iOut}{- path to save files}
}
\description{
Downloads data from ENCODE project
}
\examples{
\dontrun{
   encodeDownloader('bone chip',
                   'experiment',
                   'transcription factor',
                   'primary cell',
                   'bigWig',
                   'ChIP-seq',
                   'mm9',
                   'diretory_name_to_save_files'
                   )
}
}
\seealso{
\url{https://www.encodeproject.org/search/}

\url{https://www.encodeproject.org/help/rest-api/}
}

