% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/search.R
\name{encodeSearch}
\alias{encodeSearch}
\title{Encode search}
\usage{
encodeSearch(accession = NULL, biosample = NULL, assay = NULL,
  lab = NULL, target = NULL, description = NULL, organism = NULL)
}
\arguments{
\item{accession}{GEO sample accession Ex: ENCSR337VPD}

\item{biosample}{Example:
\tabular{lllll}{
 hepatocyte             \tab GM12878    \tab induced pluripotent stem cell
  \tab kidney    \tab Daoy                    \cr
 neural progenitor cell \tab A673       \tab bipolar spindle neuron
  \tab heart     \tab A172                    \cr
 K562                   \tab Karpas-422 \tab adipose tissue
 \tab liver     \tab LHCN-M2                 \cr
 A549                   \tab MM.1S      \tab adrenal gland
 \tab testis    \tab skeletal muscle myoblast\cr
 brain                  \tab HT1080     \tab female gonad
 \tab RPMI-7951 \tab myotube                 \cr
 spleen                 \tab SK-N-DZ    \tab lung
 \tab M059J     \tab H7-hESC                 \cr
 pancreas               \tab SK-MEL-5   \tab sigmoid colon
 \tab H4        \tab astrocyte               \cr
 HepG2                  \tab NCI-H460   \tab small intestine
 \tab SJSA1     \tab HeLa-S3
}}

\item{assay}{Example:
\tabular{ll}{
 RAMPAGE  \tab RRBS \cr
 ChIP-seq \tab FAIRE-seq \cr
 RNA-seq  \tab Repli-seq \cr
 shRNA knockdown followed by RNA-seq
 \tab DNA methylation profiling by array assay            \cr
 DNase-seq \tab RIP-seq \cr
 iCLIP     \tab Switchgear \cr
 ChIA-PET  \tab CAGE   \cr
 MeDIP-seq assay \tab RIP-chip   \cr
 MRE-seq         \tab Repli-chip \cr
 RNA profiling by array assay
  \tab whole-genome shotgun bisulfite sequencing           \cr
 RNA-PET         \tab MNase-seq  \cr
 5C              \tab DNA-PET    \cr
 comparative genomic hybridization by array
 \tab protein sequencing by tandem mass spectrometry assay
}}

\item{lab}{Example:
\tabular{lllll}{
Thomas Gingeras, CSHL    \tab John Stamatoyannopoulos, UW
\tab Kevin Struhl, HMS    \tab Gregory Crawford, Duke
 \tab David Gilbert, FSU      \cr
Kevin White, UChicago    \tab Barbara Wold, Caltech
\tab Peggy Farnham, USC   \tab Morgan Giddings, UNC
\tab Ali Mortazavi, UCI      \cr
Michael Snyder, Stanford \tab Gene Yeo, UCSD
\tab Job Dekker, UMass    \tab Scott Tenenbaum, SUNY-Albany
\tab Ross Hardison, PennState\cr
Brenton Graveley, UConn  \tab Bradley Bernstein, Broad
\tab Vishwanath Iyer, UTA \tab Piero Carninci, RIKEN
\tab Bing Ren, UCSD          \cr
Richard Myers, HAIB      \tab Yijun Ruan, GIS
\tab Jason Lieb, UNC      \tab Sherman Weissman, Yale \tab Joe Ecker, Salk
}}

\item{target}{Target Example:
\tabular{lllll}{
 Control            \tab SERBP1 \tab FAM120A \tab DDX27  \tab RPLP0 \cr
 rabbit-IgG-control \tab RRP9   \tab EIF3D   \tab CSTF2  \tab RPL23A\cr
 NUSAP1             \tab RPS19  \tab EFTUD2  \tab BCLAF1 \tab RCC2  \cr
 XRN2               \tab PRPF8  \tab EEF2    \tab BCCIP  \tab RBM27 \cr
 UCHL5              \tab PES1   \tab DDX55   \tab UPF2   \tab PSIP1 \cr
 TFIP11             \tab PA2G4  \tab DDX52   \tab TROVE2 \tab PKM   \cr
 SUPV3L1            \tab NONO   \tab DDX51   \tab RPS5   \tab PHF6  \cr
 SRP68              \tab GEMIN5 \tab DDX28   \tab RPS2   \tab
 Non-specific target control
}}

\item{description}{Description of the sample}

\item{organism}{The organism that the sample belongs to
\tabular{l}{
Homo sapiens \cr Mus musculus \cr Drosophila melanogaster
}}
}
\value{
Dataframe with the query result
}
\description{
Searches in the encode database
}
\examples{
# all database
query <- encodeSearch()

query <- encodeSearch(biosample = "brain" )

query <- encodeSearch(biosample = c("brain","spleen") )

query <- encodeSearch(biosample = "brain", assay = "RNA-seq" )

query <- encodeSearch(accession = "ENCSR554PHF" )

query <- encodeSearch(organism = "Mus musculus" )
}

