% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roadmap.R
\name{geoDownloader}
\alias{geoDownloader}
\title{Download data from GEO database}
\usage{
geoDownloader(query)
}
\arguments{
\item{query}{- GEO query -  http://www.ncbi.nlm.nih.gov/gds/advanced}
}
\value{
Download data
}
\description{
Download data from GEO database using reutils library
             Input should be a GEO query, you can you this site
             to create the query  http://www.ncbi.nlm.nih.gov/gds/advanced
}
\examples{
   geoDownloader ("(h1[All Fields] AND RRBS[All Fields])")
 query <- "((GSM956006) AND gsm[Filter]) AND roadmap epigenomics[Project]"
 query <- "((h1 cell ) AND gsm[Filter]) AND roadmap epigenomics[Project]"
}
\seealso{
\url{http://www.ncbi.nlm.nih.gov/geo/info/download.html}

\url{http://www.ncbi.nlm.nih.gov/geo/info/geo_paccess.html}

\url{http://www.ncbi.nlm.nih.gov/books/NBK25499/#chapter4.ESearch}
}

