% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/roadmap.R
\name{geoDownloaderLinks}
\alias{geoDownloaderLinks}
\title{Download a files from a list o ftp directory and uncompress it
Download a files from a list o ftp directory and uncompress it}
\usage{
geoDownloaderLinks(iFTPs, iOut, gui = FALSE)

geoDownloaderLinks(iFTPs, iOut, gui = FALSE)
}
\arguments{
\item{iFTPs:}{list of ftp directory adress
iOut - folder where files will be downloaded}

\item{iFTPs:}{list of ftp directory adress
iOut - folder where files will be downloaded}
}
\value{
download geo links
}
\description{
Download a files from a list o ftp directory and uncompress it
Download a files from a list o ftp directory and uncompress it

Download a files from a list o ftp directory and uncompress it
Download a files from a list o ftp directory and uncompress it
}
\examples{
geoDownloaderLinks (
    c('ftp://ftp.ncbi.nlm.nih.gov/geo/samples/GSM409nnn/GSM409307/suppl/',
      'ftp://ftp.ncbi.nlm.nih.gov/geo/samples/GSM409nnn/GSM409312/suppl/'
      ),
    'path_to_folder_GSM409307'
  )
\dontrun{
 geoDownloaderLinks (
    c('ftp://ftp.ncbi.nlm.nih.gov/geo/samples/GSM409nnn/GSM409307/suppl/',
      'ftp://ftp.ncbi.nlm.nih.gov/geo/samples/GSM409nnn/GSM409312/suppl/'
      ),
    'path_to_folder_GSM409307'
  )
}
}
\keyword{internal}

