% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{get.obj}
\alias{get.obj}
\title{Get object from inside the package}
\usage{
get.obj(obj)
}
\value{
Return object from package
}
\description{
Get object from inside the package
}
\examples{
maf.files <- get.obj("maf.files")
}
\keyword{internal}

