% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/biOmics.R
\docType{data}
\name{platforms}
\alias{platforms}
\title{Platforms of roadmap, encode and TCGA mapped into groups.}
\format{A data frame with 123 rows and 2 variables}
\description{
Platforms of roadmap, encode and TCGA mapped into groups.
}
\keyword{internal}

