% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/prime.R
\name{prime}
\alias{prime}
\title{Verify if a number is prime}
\usage{
prime(x)
}
\arguments{
\item{x}{A number}
}
\value{
TRUE if \code{x} is prime, FALSE otherwise
}
\description{
\code{prime} verify if a number is prime
}
\examples{
prime(4)
prime(13)
\dontrun{
   prime('a')
}
}
\seealso{
\url{http://en.wikipedia.org/wiki/Prime_number}
}

