% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/biOmics.R
\docType{data}
\name{roadmap.db}
\alias{roadmap.db}
\title{Roadmap metadata database.
The data set contains the following fields:
\itemize{
  \item X..GEO.Accession Geo Acession number
  \item Sample.Name	Sample name
  \item Experiment	 Experiment name
  \item NA.Accession	NA acession number
  \item Center	Center name
  \item SRA.FTP	SRA file FTP path
  \item GEO.FTP	GEO files FTO path
  \item Embargo.end.date Embargo date
}}
\format{A data frame with 3135 rows and 8 variables}
\description{
Roadmap metadata database.
The data set contains the following fields:
\itemize{
  \item X..GEO.Accession Geo Acession number
  \item Sample.Name	Sample name
  \item Experiment	 Experiment name
  \item NA.Accession	NA acession number
  \item Center	Center name
  \item SRA.FTP	SRA file FTP path
  \item GEO.FTP	GEO files FTO path
  \item Embargo.end.date Embargo date
}
}
\keyword{internal}

