% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/download.R
\name{roadmap.download}
\alias{roadmap.download}
\title{Download roadmap data}
\usage{
roadmap.download(lines, type = NULL, path = ".")
}
\arguments{
\item{lines}{roadmap.search output}

\item{type}{extesion of files to be downloaded}

\item{path}{Folder to save the file}
}
\description{
Download roadmap data selected using the roadmap.search
}

