% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/download.R
\name{roadmapDownload}
\alias{roadmapDownload}
\title{Download roadmap data}
\usage{
roadmapDownload(lines, type = NULL, path = ".")
}
\arguments{
\item{lines}{roadmap.search output}

\item{type}{extesion of files to be downloaded}

\item{path}{Folder to save the file}
}
\value{
Download romapdata into path
}
\description{
Download roadmap data selected using the roadmap.search
}
\examples{
query <- roadmapSearch(sample = "H1 cell line", experiment = "RRBS")
roadmapDownload(query,type = "bed", path = "roadmap")
}

