% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/search.R
\name{roadmapSearch}
\alias{roadmapSearch}
\title{Roadmap search}
\usage{
roadmapSearch(sample = NULL, experiment = NULL)
}
\arguments{
\item{sample}{Example:
\tabular{ll}{
H1 cell line  \tab Fetal Heart\cr
IMR90 fetal lung fibroblasts Cell Line \tab Fetal Kidney\cr
Aorta \tab lung\cr
Fetal Adrenal Gland \tab CD34 cultured cells
}}

\item{experiment}{Examples:
\tabular{llll}{
DNase   \tab H2A.Z  \tab H2AK5ac \tab H2BK120ac   \cr
H2BK12ac   \tab H2BK15ac \tab H2BK20ac  \tab H2BK5ac \cr
H3K14ac  \tab H3K18ac   \tab H3K23ac  \tab H3K23me2  \cr
H3K27ac    \tab H3K27me3  \tab H3K36me3   \tab  H3K4ac   \cr
H3K4me1   \tab H3K4me2   \tab H3K4me3  \tab H4K5ac
}}
}
\value{
Dataframe with the results of the query
}
\description{
Searches in the roadmap database
   Source: https://docs.google.com/spreadsheets/d/1yikGx4MsO9Ei36b64yOy9Vb6oPC5IBGlFbYEt-N6gOM/edit#gid=14
}

