% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methylation.R
\name{starbursanalysis}
\alias{starbursanalysis}
\title{Preparing methylation data and expression data}
\usage{
starbursanalysis(met, expression)
}
\arguments{
\item{met}{methylation data}

\item{expression}{expression data}
}
\value{
Methylation table
}
\description{
Preparing methylation data and expression data
}

