% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/biOmics.R
\docType{data}
\name{systems}
\alias{systems}
\title{BTO values mapped for some selected systems.}
\format{A data frame with 17 rows and 2 variables}
\description{
BTO values mapped for some selected systems.
}
\keyword{internal}

