% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/download.R
\name{tcga.download}
\alias{tcga.download}
\title{TCGA Download}
\usage{
tcga.download(data = NULL, path = ".")
}
\arguments{
\item{data}{TCGAQUery output}

\item{path}{location of the final data saving}
}
\description{
Download data previously selected using the TCGAQuery function
}
\examples{
\dontrun{
         TCGADownload(data,'folder')
}
}
\seealso{
TCGAQuery
}

