% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/roadmap.R
\name{validateInput}
\alias{validateInput}
\title{Verify user input}
\usage{
validateInput(input = NULL, list = NULL)
}
\arguments{
\item{input}{- user input of samples}

\item{list}{- roadmap list of experiments or samples}
}
\description{
Verify user input existis in roadmap
}
\examples{
\dontrun{
 validateInput(samples,roadmap$samplesList)
}
}
\keyword{internal}

