% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/search.R
\name{biOmics.search}
\alias{biOmics.search}
\title{Searches a term in TCGA, ENCODE, ROADMAP}
\usage{
biOmics.search(term, experiment = "all")
}
\arguments{
\item{term}{Term to be searched. Example: "brain" "u87" etc.}

\item{experiment}{Experiment type
\tabular{llll}{
Microarray      \tab MiRNAMicroArray \tab RRBS    \tab DNAsequencing\cr
ExpressionArray \tab Firehose        \tab ChipSeq \tab fiveC        \cr
ExonArray       \tab DNAMethylation  \tab MRESeq  \tab RepliSeq     \cr
RNASeq          \tab miRNASeq        \tab Rampage \tab Others
}}
}
\description{
Search a term using an ontology in the TCGA, ENCODE, ROADMAP databases
}

