% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/global.R
\name{biomicsApp}
\alias{biomicsApp}
\title{biomics interface}
\usage{
biomicsApp()
}
\description{
Calls UI interface
}
\examples{
\dontrun{
   biomicsApp()
}
}
\seealso{
\url{https://www.encodeproject.org/search/}

\url{https://www.encodeproject.org/help/rest-api/}
}

