% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/global.R
\name{biOMICsApp}
\alias{biOMICsApp}
\title{biOMICs interface}
\usage{
biOMICsApp()
}
\description{
Calls UI interface
}
\examples{
\dontrun{
   biOMICsApp()
}
}
\seealso{
\url{https://www.encodeproject.org/search/}

\url{https://www.encodeproject.org/help/rest-api/}
}

