% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{biOmicsApp}
\alias{TCGAbiolinksGUI}
\alias{biOmicsApp}
\title{biOmics interface}
\usage{
biOmicsApp()

TCGAbiolinksGUI()
}
\value{
Open a connection to shiny

Open a connection to shiny
}
\description{
Calls UI interface

Calls UI interface
}
\examples{
   biOmicsApp()
   TCGAbiolinksGUI()
}
\seealso{
\url{https://www.encodeproject.org/search/}

\url{https://www.encodeproject.org/help/rest-api/}
}

