% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/download.R
\name{biOmicsDownload}
\alias{biOmicsDownload}
\title{biOmicsDownload}
\usage{
biOmicsDownload(lines = NULL, enc.file.type = NULL, rmap.file.type = NULL)
}
\arguments{
\item{lines}{biOmicsSearch output}

\item{enc.file.type}{Extension to be downloaed from encode database}

\item{rmap.file.type}{Extension to be downloaed from roadmap database}
}
\value{
Saves the file of the line files
}
\description{
Download data selected using the biOmics.search function
}
\examples{
query <- biOmicsSearch("u87")
biOmicsDownload(query, enc.file.type = "bam",rmap.file.type = "bed")
}
\seealso{
biOmicsSearch
}

