% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/encode.R
\name{encodeDownloader}
\alias{encodeDownloader}
\title{Download data from ENCODDE project}
\usage{
encodeDownloader(search, type, target, sample, ftype, assay, assembly, out)
}
\arguments{
\item{search}{- bone chip}

\item{type}{-   'experiment' 'publication' 'software' ''antibody_lot' 'web'}

\item{target}{- 'transcription factor' 'RNA binding protein'
'tag' 'histone modification'}

\item{sample}{- 'tissue' 'primary cell'}

\item{ftype}{- 'bam' 'bigWig' 'bed_broadPeak' 'broadPeak' 'fastq'}

\item{assay}{- 'ChIP-seq' 'RIP-chip' 'Repli-chip'}

\item{assembly}{- 'hg19' 'mm9'}

\item{out}{- path to save files}
}
\value{
Save encode data into folder, return data downloaded
}
\description{
Downloads data from ENCODE project
}
\examples{
encodeDownloader('bone chip',
                   'experiment',
                   'transcription factor',
                   'primary cell',
                   'bigWig',
                   'ChIP-seq',
                   'mm9',
                   'diretory_name_to_save_files'
                   )
}
\seealso{
\url{https://www.encodeproject.org/search/}

\url{https://www.encodeproject.org/help/rest-api/}
}

