% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/roadmap.R
\name{geoDownloader}
\alias{geoDownloader}
\title{Download data from GEO database}
\usage{
geoDownloader(iQuery, iOut)
}
\arguments{
\item{iQuery}{- GEO query -  http://www.ncbi.nlm.nih.gov/gds/advanced}

\item{iOut}{- path to save files}
}
\description{
Download data from GEO database using reutils library
             Input should be a GEO query, you can you this site
             to create the query  http://www.ncbi.nlm.nih.gov/gds/advanced
}
\examples{
\dontrun{
   geoDownloader (''(h1[All Fields] AND RRBS[All Fields])
                   AND roadmap epigenomics[Project] AND 'gsm'[Filter]',
                   )
 iQuery <- '((GSM956006) AND gsm[Filter]) AND roadmap epigenomics[Project]'
 iQuery <- '((h1 cell ) AND gsm[Filter]) AND roadmap epigenomics[Project]'
}
}
\seealso{
\url{http://www.ncbi.nlm.nih.gov/geo/info/download.html}

\url{http://www.ncbi.nlm.nih.gov/geo/info/geo_paccess.html}

\url{http://www.ncbi.nlm.nih.gov/books/NBK25499/#chapter4.ESearch}
}

