% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAbiolinksGUI.R, R/global.R
\docType{package}
\name{TCGAbiolinksGUI}
\alias{TCGAbiolinksGUI}
\alias{TCGAbiolinksGUI-package}
\alias{TCGAbiolinksGUI}
\title{TCGAbiolinksGUI}
\usage{
TCGAbiolinksGUI(run = TRUE)
}
\arguments{
\item{run}{Used to control the examples.}
}
\value{
Open a connection to shiny
}
\description{
A Graphical User Interface for integrative analysis of TCGA data

Calls UI interface
}
\details{
The functions you're likely to need from \pkg{TCGAbiolinksGUI} are
\code{\link{TCGAbiolinksGUI}}
}
\examples{
\dontrun{
   TCGAbiolinksGUI()
}
TCGAbiolinksGUI(run = FALSE)
}
