% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{update}
\alias{update}
\title{Update database
This function will update the biomics database}
\usage{
update()
}
\value{
Update maf files table
}
\description{
Update database
This function will update the biomics database
}
\keyword{internal}

